+++ src/components/ContactDetailsModal/ContactDetailsModal.js
@@ -38,7 +38,7 @@ import Modal from 'components/Modal';
 import ModalBox from 'components/ModalBox';
 import Spinner from 'components/Spinner';
 import TextInput from 'components/legacy/TextInput';
-import Title from 'components/Title';
+import Title from 'components/legacy/Title';
 
 // Utils
 import { addressesEqual } from 'utils/assets';
