+++ src/screens/SendToken/SendTokenTransaction.js
@@ -27,7 +27,7 @@ import t from 'translations/translate';
 // Components
 import { Container, Wrapper } from 'components/legacy/Layout';
 import { Paragraph, MediumText } from 'components/legacy/Typography';
-import Title from 'components/Title';
+import Title from 'components/legacy/Title';
 import Button from 'components/modern/Button';
 import Animation from 'components/Animation';
 import Toast from 'components/Toast';
